requirements.txt
# Core dependencies
httpx>=0.28.1,<1  # Resolves pyautogen conflict while satisfying most packages
pyautogen==0.8.7
anthropic>=0.49.0
chromadb>=0.6.3
langsmith>=0.3.18
openai>=1.68.2
ollama>=0.4.7

# Additional packages (adjust versions if needed)
pymongo>=4.5.0
beautifulsoup4>=4.12.3
lxml>=4.9.4
Pillow>=10.2.0
pyserial>=3.5
tldextract>=5.1.1

# Security/authentication
bottle-cork>=0.12.0
beaker>=1.12.0
pycryptodome>=3.20.0

requests
Flask==2.2.5
Werkzeug==2.3.0
mcp
fastapi
sse-starlette
python-dotenv
google-adk
google-cloud-secret-manager